requirements.txt
acvl_utils==0.2
batchgenerators==0.25
batchgeneratorsv2==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
connected-components-3d==3.18.0
contourpy==1.2.1
cycler==0.12.1
dicom2nifti==2.4.11
dynamic_network_architectures==0.3.1
einops==0.8.0
fft-conv-pytorch==1.2.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
graphviz==0.20.3
idna==3.7
imagecodecs==2024.6.1
imageio==2.35.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
linecache2==1.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
MedPy==0.5.2
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydicom==2.4.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gdcm==3.0.24.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
SimpleITK==2.3.1
six==1.16.0
sympy==1.13.2
threadpoolctl==3.5.0
tifffile==2024.8.10
torch @ https://download.pytorch.org/whl/cu121/torch-2.2.0%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=c441021672ebe2e5afbdb34817aa85e6d32130f94df2da9ad4cb78a9d4b81370
torchvision @ https://download.pytorch.org/whl/cu121/torchvision-0.17.0%2Bcu121-cp310-cp310-linux_x86_64.whl#sha256=e1273e9862fc821febc4c716f13849b1ffa87c0d69f6ab824356c5532c490f08
tqdm==4.66.5
traceback2==1.4.0
triton==2.2.0
typing_extensions==4.12.2
tzdata==2024.1
unittest2==1.1.0
urllib3==2.2.2
yacs==0.1.8
IPython==8.26.0
onnx==1.16.1
netron
onnx-simplifier==0.4.36
onnxoptimizer==0.3.13
onnxruntime
torchprofile==0.0.4
fvcore==0.1.5.post20221221
torchsummary==1.5.1
plotly
graphviz==0.20.3
hiddenlayer==0.2
fft_conv_pytorch==1.2.0
surface_distance==0.1
timm==1.0.15